Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -4,7 +4,7 @@
 
 int main() {
     char choice;
-    int a,b,c;
+    float a,b,c;
     scanf("%d %d %c",&a,&b,&choice);
     switch(choice)
     {
